requirements.txt
flask
openai>=1.0.0  # Updated to use the new OpenAI SDK
python-dotenv
gunicorn
flask-cors
requests
beautifulsoup4